calabi_bcs

Function to_bytes_with_capacity

pub fn to_bytes_with_capacity<T>(value: &T, capacity: usize) -> Result<Vec<u8>>
where
    T: ?Sized + TBCSSerialize,

Same as to_bytes but pre-allocates the output buffer with the given capacity.

This can improve performance when you have a good estimate of the serialized size, as it avoids reallocations during serialization.

Errors

Returns an error if any error condition from to_bytes occurs.

Examples

use calabi_bcs::{to_bytes_with_capacity, BcsSerialize};
use serde::Serialize;

#[derive(Serialize, BcsSerialize)]
struct Data {
    values: Vec<u64>,
}

let data = Data { values: vec![1, 2, 3, 4, 5] };
// Pre-allocate for length prefix (1 byte) + 5 u64s (40 bytes)
let bytes = to_bytes_with_capacity(&data, 41).unwrap();